refactor sendTweet in twitter api
